feat: remove 'use client' directive from index files and refactor withIoc function
