+++ packages/gene-apollo/src/index.ts
@@ -1,2 +1,1 @@
-'use client';
 export * from './apollo';
