+++ packages/gene-next/src/index.ts
@@ -1,4 +1,3 @@
-'use client';
 export * from './api-routes';
 export * from './routing';
 export * from './translations';
