+++ packages/gene/src/ioc/react.ioc.tsx
@@ -1,3 +1,4 @@
+'use client';
 import React, { PropsWithChildren, useContext, useMemo } from 'react';
 
 import { interfaces } from 'inversify';
@@ -106,20 +107,4 @@ export function useCallbackInjection<T, R = void>(
   } catch {
     return fallback;
   }
-}
-
-export function withIoc<Props extends Record<string, any>>(
-  getContainer: (props?: Props) => interfaces.Container
-) {
-  return (Page: React.ComponentType<Props>) => {
-    return (props: Props) => {
-      const container = getContainer(props);
-
-      return (
-        <Provider container={container}>
-          <Page {...props} />
-        </Provider>
-      );
-    };
-  };
 }
